modification run + calcul objectif

+++ include/Random.hpp
@@ -9,6 +9,7 @@
 #include <fstream>
 #include <time.h>
 #include <algorithm>
+#include <random>
 #include "Interface.hpp"
 
 
@@ -17,13 +18,13 @@ class Random
 {
     public:
         static void randomize();
-        static long aleatoire(int borne);
+        static long aleatoire(long borne);
         static void melangerAleatoirementInterfaces(std::vector<Interface>& interfaces);
-
 
     protected:
 
     private:
+        
        
 };
